Remove hardcoded Coveo content

+++ build/coveo-for-sitecore-tools/Generate-Coveo-Assets-Folders.ps1
@@ -9,17 +9,32 @@ $parentFolderPath = (get-item $PSScriptRoot).parent.FullName
 $coveoAssetsFolderPath = "$parentFolderPath\windows\$SitecoreVersion\modules\coveo-assets"
 $coveoSxaAssetsFolderPath = "$parentFolderPath\windows\$SitecoreVersion\modules\coveo-sxa-assets"
 
-Remove-Item $coveoAssetsFolderPath\*.zip | Out-Null
-Remove-Item $coveoSxaAssetsFolderPath\*.zip | Out-Null
+if (Test-Path -Path $coveoAssetsFolderPath\*.zip -PathType Leaf)
+{
+    Remove-Item $coveoAssetsFolderPath\*.zip | Out-Null
+}
+
+if (Test-Path -Path $coveoSxaAssetsFolderPath\*.zip -PathType Leaf)
+{
+    Remove-Item $coveoSxaAssetsFolderPath\*.zip | Out-Null
+}
+
 
 #Creating required folders for Coveo assets
 New-Item -Path $coveoAssetsFolderPath -ItemType directory -Force | Out-Null
 New-Item -Path "$coveoAssetsFolderPath\tools" -ItemType directory -Force | Out-Null
 
-#Creating required folders for Coveo SXA assets
-New-Item -Path $coveoSxaAssetsFolderPath -ItemType directory -Force | Out-Null
-New-Item -Path "$coveoSxaAssetsFolderPath\tools" -ItemType directory -Force | Out-Null
+if ($IncludeSxa)
+{
+    #Creating required folders for Coveo SXA assets
+    New-Item -Path $coveoSxaAssetsFolderPath -ItemType directory -Force | Out-Null
+    New-Item -Path "$coveoSxaAssetsFolderPath\tools" -ItemType directory -Force | Out-Null
+}
 
 #Copying the Extract-Resource.ps1 script into the tools folders
 Copy-Item "$PSScriptRoot\Extract-Resources.ps1" -Destination "$coveoAssetsFolderPath\tools" | Out-Null
-Copy-Item "$PSScriptRoot\Extract-Resources.ps1" -Destination "$coveoSxaAssetsFolderPath\tools" | Out-Null
+
+if ($IncludeSxa)
+{
+    Copy-Item "$PSScriptRoot\Extract-Resources.ps1" -Destination "$coveoSxaAssetsFolderPath\tools" | Out-Null
+}
